Split test_cite_command from test_citations
