+++ manubot/cite/tests/test_cite_command.py
@@ -1,0 +1,87 @@
+import json
+import pathlib
+import shutil
+import subprocess
+
+import pytest
+
+
+def test_cite_command_empty():
+    process = subprocess.run(
+        ['manubot', 'cite'],
+        stdout=subprocess.PIPE,
+        stderr=subprocess.PIPE,
+        universal_newlines=True,
+    )
+    print(process.stderr)
+    assert process.returncode == 2
+    assert 'the following arguments are required: citations' in process.stderr
+
+
+def test_cite_command_stdout():
+    process = subprocess.run(
+        ['manubot', 'cite', 'arxiv:1806.05726v1'],
+        stdout=subprocess.PIPE,
+        stderr=subprocess.PIPE,
+        universal_newlines=True,
+    )
+    print(process.stderr)
+    assert process.returncode == 0
+    csl, = json.loads(process.stdout)
+    assert csl['URL'] == 'https://arxiv.org/abs/1806.05726v1'
+
+
+def test_cite_command_file(tmpdir):
+    path = pathlib.Path(tmpdir) / 'csl-items.json'
+    process = subprocess.run(
+        ['manubot', 'cite', '--output', str(path), 'arxiv:1806.05726v1'],
+        stdout=subprocess.PIPE,
+        stderr=subprocess.PIPE,
+    )
+    print(process.stderr.decode())
+    assert process.returncode == 0
+    with path.open() as read_file:
+        csl, = json.load(read_file)
+    assert csl['URL'] == 'https://arxiv.org/abs/1806.05726v1'
+
+
+@pytest.mark.parametrize(['args', 'expected'], [
+    ([], 'references-plain.txt'),
+    (['--format', 'plain'], 'references-plain.txt'),
+    (['--format', 'markdown'], 'references-markdown.md'),
+    (['--format', 'html'], 'references-html.html'),
+    (['--format', 'jats'], 'references-jats.xml'),
+], ids=['no-args', '--format=plain', '--format=markdown', '--format=html', '--format=jats'])
+@pytest.mark.skipif(
+    not shutil.which('pandoc'),
+    reason='pandoc installation not found on system'
+)
+@pytest.mark.skipif(
+    not shutil.which('pandoc-citeproc'),
+    reason='pandoc-citeproc installation not found on system'
+)
+def test_cite_command_render_stdout(args, expected):
+    """
+    Note that this test may fail if the Pandoc version is not recent enough to
+    support --lua-filter (introduced in pandoc 2.0) or URLs for --csl.
+    """
+    expected = (
+        pathlib.Path(__file__).parent
+        .joinpath('cite-command-rendered', expected)
+        .read_text()
+    )
+    args = [
+        'manubot', 'cite', '--render',
+        '--csl', 'https://github.com/greenelab/manubot-rootstock/raw/e83e51dcd89256403bb787c3d9a46e4ee8d04a9e/build/assets/style.csl',
+        'arxiv:1806.05726v1', 'doi:10.6084/m9.figshare.5346577.v1', 'pmid:29618526',
+    ] + args
+    process = subprocess.run(
+        args,
+        stdout=subprocess.PIPE,
+        stderr=subprocess.PIPE,
+        universal_newlines=True,
+    )
+    print(' '.join(process.args))
+    print(process.stdout)
+    print(process.stderr)
+    assert process.stdout == expected
